requirements.txt
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2
django-cors-headers==2.5.3
django-extensions==2.1.6
django-heroku==0.3.1
django-localflavor==2.1
django-rest-auth==0.9.5
django-rest-passwordreset==1.0.0
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
Pillow==5.3.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
PyJWT==1.7.1
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2018.4
reportlab==3.5.23
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.3
webencodings==0.5.1
whitenoise==4.1.2
xhtml2pdf==0.2.3
